Moved Resources in a separate namespace - to keep module clean structured
